Moving templates to the top of modal, cleaning up
